+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Tasks are no longer executed in parallel [#606](https://github.com/CCI-Tools/cate/issues/606).
 * Improve error messages and handling [#393](https://github.com/CCI-Tools/cate/issues/393),
   introduced new error type `cate.core.types.ValidationError` for special treatment in the GUI.
+* Make Cate HTTP User-Agent distinguishable [#510](https://github.com/CCI-Tools/cate/issues/510).
 * Fixed broken WebAPI invocation from CLI.
 
 ## Version 2.0.0.dev8