+++ CHANGES.md
@@ -1,7 +1,7 @@
 ## Version 2.0.0.dev10 (in development)
 
 * Fixed a bug with animation generation [#585](https://github.com/CCI-Tools/cate/issues/585)
-* Upgrade to using newer xarray version after an upstream bugfix []()
+* Upgrade to using newer xarray version after an upstream bugfix [#579](https://github.com/CCI-Tools/cate/issues/579)
 
 ## Version 2.0.0.dev9